backend/requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据验证和序列化
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.8

# 文件处理
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP客户端
httpx==0.25.2
aiofiles==23.2.1

# 日志
loguru==0.7.2

# 环境变量
python-dotenv==1.0.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0